+++ src/main/java/com/sparta/settlementservice/batch/config/BatchConfig.java
@@ -1,28 +1,30 @@
 package com.sparta.settlementservice.batch.config;
 
-import org.springframework.batch.core.configuration.JobRegistry;
-import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
-import org.springframework.batch.core.explore.JobExplorer;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.launch.support.SimpleJobOperator;
-import org.springframework.batch.core.repository.JobRepository;
+
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.partition.PartitionHandler;
+import org.springframework.batch.core.partition.support.TaskExecutorPartitionHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.core.task.TaskExecutor;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 @Configuration
-@EnableBatchProcessing
 public class BatchConfig {
-
+    //TaskExecutor 빈이 하나로만 정의되어 있지만, 파티셔닝을 이용해서
+    // 병렬 처리를 하도록 설정하는 것이므로 TaskExecutor가 여러 개로 인식되는 문제는 @Primary 어노테이션으로 해결이 가능
+    @Primary
     @Bean
-    public SimpleJobOperator jobOperator(JobLauncher jobLauncher,
-                                         JobRepository jobRepository,
-                                         JobRegistry jobRegistry,
-                                         JobExplorer jobExplorer) {
-        SimpleJobOperator jobOperator = new SimpleJobOperator();
-        jobOperator.setJobLauncher(jobLauncher);
-        jobOperator.setJobRepository(jobRepository);
-        jobOperator.setJobRegistry(jobRegistry);
-        jobOperator.setJobExplorer(jobExplorer);
-        return jobOperator;
+    public TaskExecutor taskExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(4);  // 최소 4개의 스레드
+        executor.setMaxPoolSize(8);   // 최대 8개의 스레드
+        executor.setQueueCapacity(100);
+        executor.setThreadNamePrefix("batch-executor-");
+        executor.initialize();
+        return executor;
     }
+
+
 }
