--- src/main/java/com/sparta/settlementservice/batch/controller/MainController.java
@@ -1,4 +1,0 @@
-package com.sparta.settlementservice.batch.controller;
-
-public class MainController {
-}
