+++ src/main/java/com/sparta/settlementservice/batch/settlementbatch/DailyStaticBatch.java
@@ -33,28 +33,25 @@ public class DailyStaticBatch {
 
     private final JobRepository jobRepository;
     private final PlatformTransactionManager platformTransactionManager;
-    private final DailyViewPlaytimeRepository dailyViewPlaytimeRepository;
-    private final DailyVideoViewRepository dailyVideoViewRepository;
     private final DailyViewPlaytimeJdbcRepository dailyViewPlaytimeJdbcRepository;
+    private final TaskExecutor taskExecutor;
 
     @Autowired
     public DailyStaticBatch(JobRepository jobRepository,
                             PlatformTransactionManager platformTransactionManager,
-                            DailyViewPlaytimeRepository dailyViewPlaytimeRepository,
-                            DailyVideoViewRepository dailyVideoViewRepository,
-                            DailyViewPlaytimeJdbcRepository dailyViewPlaytimeJdbcRepository
+                            DailyViewPlaytimeJdbcRepository dailyViewPlaytimeJdbcRepository,
+                            TaskExecutor taskExecutor
+
     ) {
         this.jobRepository = jobRepository;
         this.platformTransactionManager = platformTransactionManager;
-        this.dailyViewPlaytimeRepository = dailyViewPlaytimeRepository;
-        this.dailyVideoViewRepository = dailyVideoViewRepository;
         this.dailyViewPlaytimeJdbcRepository = dailyViewPlaytimeJdbcRepository;
+        this.taskExecutor = taskExecutor;
     }
 
     // Job 구성
     @Bean
     public Job dailyStatisticsJob() {
-        System.out.println("[Job] Starting 'dailyStatisticsJob'");
         return new JobBuilder("dailyStatisticsJob", jobRepository)
                 .start(partitionedDailyStatisticsStep()) // 파티션 Step 실행
                 .build();
@@ -63,30 +60,15 @@ public class DailyStaticBatch {
 
     @Bean
     public Partitioner videoIdPartitioner() {
-        System.out.println("[Partitioner] Initialized VideoIdPartitioner");
         return new VideoIdPartitioner(); // 너가 작성한 Partitioner
     }
 
-    //TaskExecutor 빈이 하나로만 정의되어 있지만, 파티셔닝을 이용해서
-    // 병렬 처리를 하도록 설정하는 것이므로 TaskExecutor가 여러 개로 인식되는 문제는 @Primary 어노테이션으로 해결이 가능
-    @Primary
     @Bean
-    public TaskExecutor taskExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(4); // 최소 4개의 스레드
-        executor.setMaxPoolSize(8);  // 최대 8개의 스레드
-        executor.setQueueCapacity(100); // 대기 큐의 크기
-        executor.setThreadNamePrefix("batch-executor-");
-        executor.initialize();
-        return executor;
-    }
-
-    @Bean
-    public PartitionHandler partitionHandler(TaskExecutor taskExecutor) {
+    public PartitionHandler partitionHandler(TaskExecutor taskExecutor, Step dailyStatisticsStep) {
         TaskExecutorPartitionHandler partitionHandler = new TaskExecutorPartitionHandler();
-        partitionHandler.setTaskExecutor(taskExecutor); // 병렬 처리할 TaskExecutor 지정
-        partitionHandler.setStep(dailyStatisticsStep()); // 각 파티션에 대해 실행할 Step 지정
-        partitionHandler.setGridSize(4); // 병렬 처리 수 (4개의 파티션)
+        partitionHandler.setTaskExecutor(taskExecutor); // 멀티스레드 실행
+        partitionHandler.setStep(dailyStatisticsStep);  // 실행할 Step 지정
+        partitionHandler.setGridSize(4); // 병렬 실행 개수 (4개)
         return partitionHandler;
     }
 
@@ -94,12 +76,11 @@ public class DailyStaticBatch {
     @Bean
     @JobScope
     public Step partitionedDailyStatisticsStep() {
-        System.out.println("[Step] Starting 'partitionedDailyStatisticsStep'");
         return new StepBuilder("partitionedDailyStatisticsStep", jobRepository)
                 .partitioner("dailyStatisticsStep", videoIdPartitioner()) // 파티션 설정
-                .step(dailyStatisticsStep()) // 기존 Step 재사용
-                .partitionHandler(partitionHandler(taskExecutor())) // 병렬 처리 핸들러 추가
-                .gridSize(4) // 병렬 처리 수 (4개의 파티션)
+                .step(dailyStatisticsStep()) // 실행할 step 지정
+                .partitionHandler(partitionHandler(taskExecutor, dailyStatisticsStep())) // 병렬 실행
+                .gridSize(4) // 병렬 처리 수 지정
                 .build();
     }
