+++ src/main/java/com/sparta/settlementservice/user/oauth2/CustomSuccessHandler.java
@@ -15,7 +15,6 @@ import java.io.IOException;
 import java.util.Collection;
 import java.util.Iterator;
 
-;
 
 @Component
 public class CustomSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
